Refactor: Moved utils related to ROC plots to plot_util.py.
